So... Added NUnit.Engine functionality to SchedulerUnitTest.cs (base class) broke the ScriptedSubsystemCS.cs Assembly class from a .NET depency bug from hell. Not going to worry about that, removed it and all of its calls.
Going back to focus on the test methods. Unfortunately, somehow, GenerateExhaustiveSystemScheduleCombosTest.cs is running but not passing again... Like. What?? It was fine a second ago. This has been such a journey. At least my Test Explorer is working again in VSCode (I literally thing the NUnit.Engine linrary refernce in the HSFScheduler.csproj file that was added broke that too. Going to treat it like anyhting good and well in the universe and not ask anymore questions).

Back to debugging Tests and trudging on with more test development. Wish me luck <3