This commit the 'SystemSchedule(StateHistory oldStates, Stack<Access> newAccessStack, double currentTime)' constructor was updated to accurately reflect HSF handling of Event/Task Start/End time. See notes in comments or explanation below:
Explanation:
'Event' in HSF is taken to be the fundamental timestep windows the program/scheduler traverses.
'Task' in HSF is taken to be the system's asset actions that are taken WITHIN Events.
Thus, Tasks have to be fully envleoped within an event. Task Start time is set to earliest access within a given Event, and TaskEnd time is set to the latest accessible time within a given Event. This may be the Event Start/End times themselves.
Fuirthermore, although Access Scripting is not live curently, and the default Access-combo generation uses Access across the entire Simulation time, Exception handling implemented in the Constructor to ensure that invalid Access times do not enter main scheduler/potential schedules list.

This commit should be merged into <main> branch of Horizon. At least the SystemSchedules.cs file, and it would be useful to include all ScriptedSubsystemCS.cs updates too, as that is now working. Ideally everything, but most else is within the test/ directory where the HSFSchedulerUnitTest NUnit test framework is actively being developed.
Godspeed :)