Halfway through refactoring CanAddTasksUnitTest.cs to be less complex (and repeat function logic in test) to being more 'unit'-level.... In this commit:  1. CanAddTasksUnitTest.cs is working (complexly). Integration tetss included  2. AllStates.timesCompletedTaskUnitTest.cs has been generated, and is passing, but need to be checked...
More importantly:
timesCompletedTask() is not used in the current verison of CanAddTasks. This will be updated to be used again. At which point CanAddTasks will have more of a unit stlye test, with the addition of timesCompletedTasks successfully being refactored to account for tass ACROSS assets in AllStates.Events. Then, TimeDeconfliciton will serve as the unit test integraiton step...