Alright we made itgit add . CanAddTasksUnitTest.cs & timesCompleteTaskUnitTest.cs are complete and passing! Notably, a README.md file was added for both for documentation's sake.
Also, CanAddTasksUnitTest.cs was decided to include the complex pattern matching that it is currently supposed to validate. Any changes to how it is decided that CanAddTass() should be able to pass (i.e. if One Asset is upgraded to be allowed to do nothing while another one executes a task (in order to stay within good time/less than MaxTimesCompletedTask)), then this test suite would need to be revisited (Along with the documentation) to see which ones will need to be changed to support the new logic. Estimation is that The last (most complex test) in CanAddTasks will have to be written, and the other two complex ones changed (to reflect the difference in combinatorics). Oh-- and the combinatorics would have to be updated to support 0s. (But that 'test' is just a visualization anyway.

It is onto CropSchedules and merge... yeehaw