First iteration of the scheduling hashing output and analysis. This commit contains the first pass version of the ScheduleDataCapture class which does capture data... Hashing was foundto be better for quick verificaiton.
That said, this hashing analysis was still young in that it was taking the output of the program or test systemSchedules list (after it had passed through multiple default .NET .Sort() => calls), and given that this is apparently unstable (does not always return same values) in .NET, the hashing was sometimes off.

NExt commit: created a deterministic sorting replacement AND a blockchain-style ScheduleHashHistory to preserve order/permutaiton uniquesness.