Removing personal notes/documnetation on main to be just kept on my branch jebeals-scheduler.
Also, notably, I wrestlted with git to get jebeals-scheduler to the same state as jebeals-scheduler-merge to keep a linear history. Last merge was with the now-fast-forwarded jebeals-scheduler branch that I will continue to work off of. jebeals-scheduler-merge and jebeals-scheduler should now effectively be the same. The commit and merge history will reflect this (I think) given fast-forwarding was used.

The last merge:
-Created Schedule hashing and StateHistory hasing (SchedulerHashHistory and StateHistoryHashHistory) to identify unique scheudler/state states for easier comparison and evaluation of reproductive algoriothmic performance. These have an option to be turned off (Extra computaiton) in the program. They are primarily infomrational and will be sued for the Scheduler unit testing.